+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `postTarget` as alternative to `onFormSubmitted` for `Form`
 - [MINOR] Updated `Stack` to have responsive props for size
 - [MINOR] Updated `Grid.Item` to be flex by default
+- [MINOR] Added `shouldSkipRenderingWhenCollapsed` to `TitledCollapsibleBox`
+- [MINOR] Updated `Checkbox` to not render check when not checked
 
 ### Removed